// web/src/api/generated/models/PrivateCollectionDetail.ts
export interface PrivateCollectionDetail {
    /**
     * 
     * @type {ResourceCapabilities}
     * @memberof PrivateCollectionDetail
     */
    capabilities: ResourceCapabilities;
    /**
     * 
     * @type {Date}
     * @memberof PrivateCollectionDetail
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof PrivateCollectionDetail
     */
    id: string;
    /**
     * 
     * @type {Array<PrivateCollectionItemDetail>}
     * @memberof PrivateCollectionDetail
     */
    items: Array<PrivateCollectionItemDetail>;
    /**
     * 
     * @type {string}
     * @memberof PrivateCollectionDetail
     */
    ownerId: string;
    /**
     * 
     * @type {string}
     * @memberof PrivateCollectionDetail
     */
    spaceId: string;
    /**
     * 
     * @type {string}
     * @memberof PrivateCollectionDetail
     */
    title: string;
    /**
     * 
     * @type {Date}
     * @memberof PrivateCollectionDetail
     */
    updatedAt: Date;
    /**
     * 
     * @type {number}
     * @memberof PrivateCollectionDetail
     */
    version: number;
}